+++ src/Libraries/SmartStore.Core/Domain/Shipping/ShippingMethod.cs
@@ -2,6 +2,7 @@ using System.Collections.Generic;
 using System.Runtime.Serialization;
 using SmartStore.Core.Domain.Directory;
 using SmartStore.Core.Domain.Localization;
+using SmartStore.Core.Domain.Stores;
 
 namespace SmartStore.Core.Domain.Shipping
 {
@@ -9,8 +10,8 @@ namespace SmartStore.Core.Domain.Shipping
 	/// Represents a shipping method (used for offline shipping rate computation methods)
 	/// </summary>
 	[DataContract]
-	public partial class ShippingMethod : BaseEntity, ILocalizedEntity
-    {
+	public partial class ShippingMethod : BaseEntity, ILocalizedEntity, IStoreMappingSupported
+	{
         private ICollection<Country> _restrictedCountries;
 
         /// <summary>
@@ -31,12 +32,21 @@ namespace SmartStore.Core.Domain.Shipping
 		[DataMember]
 		public int DisplayOrder { get; set; }
 
+		/// <summary>
+		/// Gets or sets whether to ignore charges
+		/// </summary>
 		[DataMember]
 		public bool IgnoreCharges { get; set; }
 
-        /// <summary>
-        /// Gets or sets the restricted countries
-        /// </summary>
+		/// <summary>
+		/// Gets or sets a value indicating whether the entity is limited/restricted to certain stores
+		/// </summary>
+		[DataMember]
+		public bool LimitedToStores { get; set; }
+
+		/// <summary>
+		/// Gets or sets the restricted countries
+		/// </summary>
 		[DataMember]
 		public virtual ICollection<Country> RestrictedCountries
         {