+++ src/Libraries/SmartStore.Data/Migrations/201802081135066_ShippingMethodMultistore.cs
@@ -1,0 +1,18 @@
+namespace SmartStore.Data.Migrations
+{
+    using System;
+    using System.Data.Entity.Migrations;
+    
+    public partial class ShippingMethodMultistore : DbMigration
+    {
+        public override void Up()
+        {
+            AddColumn("dbo.ShippingMethod", "LimitedToStores", c => c.Boolean(nullable: false));
+        }
+        
+        public override void Down()
+        {
+            DropColumn("dbo.ShippingMethod", "LimitedToStores");
+        }
+    }
+}