+++ src/Presentation/SmartStore.Web/Controllers/CheckoutController.cs
@@ -185,20 +185,21 @@ namespace SmartStore.Web.Controllers
 		{
 			var model = new CheckoutShippingMethodModel();
 
-			var getShippingOptionResponse = _shippingService.GetShippingOptions(cart, _workContext.CurrentCustomer.ShippingAddress, "", _storeContext.CurrentStore.Id);
+			var store = _storeContext.CurrentStore;
+			var customer = _workContext.CurrentCustomer;
+			var getShippingOptionResponse = _shippingService.GetShippingOptions(cart, customer.ShippingAddress, "", store.Id);
 
 			if (getShippingOptionResponse.Success)
 			{
-				//performance optimization. cache returned shipping options.
-				//we'll use them later (after a customer has selected an option).
-				_genericAttributeService.SaveAttribute(_workContext.CurrentCustomer,
-					SystemCustomerAttributeNames.OfferedShippingOptions, getShippingOptionResponse.ShippingOptions, _storeContext.CurrentStore.Id);
+				// Performance optimization. cache returned shipping options.
+				// We'll use them later (after a customer has selected an option).
+				_genericAttributeService.SaveAttribute(customer, SystemCustomerAttributeNames.OfferedShippingOptions, getShippingOptionResponse.ShippingOptions, store.Id);
 
-				var shippingMethods = _shippingService.GetAllShippingMethods();
+				var shippingMethods = _shippingService.GetAllShippingMethods(null, store.Id);
 
 				foreach (var shippingOption in getShippingOptionResponse.ShippingOptions)
 				{
-					var soModel = new CheckoutShippingMethodModel.ShippingMethodModel()
+					var soModel = new CheckoutShippingMethodModel.ShippingMethodModel
 					{
 						ShippingMethodId = shippingOption.ShippingMethodId,
 						Name = shippingOption.Name,
@@ -212,12 +213,10 @@ namespace SmartStore.Web.Controllers
 						soModel.BrandUrl = _pluginMediator.GetBrandImageUrl(srcmProvider.Metadata);
 					}
 
-					//adjust rate
+					// Adjust rate.
 					Discount appliedDiscount = null;
-					var shippingTotal = _orderTotalCalculationService.AdjustShippingRate(
-						shippingOption.Rate, cart, shippingOption.Name, shippingMethods, out appliedDiscount);
-
-					decimal rateBase = _taxService.GetShippingPrice(shippingTotal, _workContext.CurrentCustomer);
+					var shippingTotal = _orderTotalCalculationService.AdjustShippingRate(shippingOption.Rate, cart, shippingOption.Name, shippingMethods, out appliedDiscount);
+					decimal rateBase = _taxService.GetShippingPrice(shippingTotal, customer);
 					decimal rate = _currencyService.ConvertFromPrimaryStoreCurrency(rateBase, _workContext.WorkingCurrency);
 					soModel.FeeRaw = rate;
 					soModel.Fee = _priceFormatter.FormatShippingPrice(rate, true);
@@ -225,30 +224,38 @@ namespace SmartStore.Web.Controllers
 					model.ShippingMethods.Add(soModel);
 				}
 
-				//find a selected (previously) shipping method
-				var selectedShippingOption = _workContext.CurrentCustomer.GetAttribute<ShippingOption>(SystemCustomerAttributeNames.SelectedShippingOption, _storeContext.CurrentStore.Id);
+				// Find a selected (previously) shipping method.
+				var selectedShippingOption = customer.GetAttribute<ShippingOption>(SystemCustomerAttributeNames.SelectedShippingOption, store.Id);
 				if (selectedShippingOption != null)
 				{
-					var shippingOptionToSelect = model.ShippingMethods.ToList()
+					var shippingOptionToSelect = model.ShippingMethods
+						.ToList()
 						.Find(so => !String.IsNullOrEmpty(so.Name) && so.Name.Equals(selectedShippingOption.Name, StringComparison.InvariantCultureIgnoreCase) &&
 						!String.IsNullOrEmpty(so.ShippingRateComputationMethodSystemName) &&
 						so.ShippingRateComputationMethodSystemName.Equals(selectedShippingOption.ShippingRateComputationMethodSystemName, StringComparison.InvariantCultureIgnoreCase));
 
 					if (shippingOptionToSelect != null)
+					{
 						shippingOptionToSelect.Selected = true;
+					}
 				}
-				//if no option has been selected, let's do it for the first one
+
+				// If no option has been selected, let's do it for the first one.
 				if (model.ShippingMethods.Where(so => so.Selected).FirstOrDefault() == null)
 				{
 					var shippingOptionToSelect = model.ShippingMethods.FirstOrDefault();
 					if (shippingOptionToSelect != null)
+					{
 						shippingOptionToSelect.Selected = true;
+					}
 				}
 			}
 			else
 			{
 				foreach (var error in getShippingOptionResponse.Errors)
+				{
 					model.Warnings.Add(error);
+				}
 			}
 
 			return model;